pyproject.toml
[project]
name = "7tv-to-sticker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.3",
    "fastapi>=0.128.0",
    "jinja2>=3.1.6",
    "python-multipart>=0.0.21",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
    "ffmpeg-python>=0.2.0",
    "gql>=4.0.0",
    "requests-toolbelt>=1.0.0",
    "pydantic-settings>=2.12.0",
    "pre-commit>=4.5.1",
    "pyupgrade>=3.21.2",
]
